pyproject.toml
[project]
name = "claim-submission-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "zeep>=4.3.1",
    "xmltodict>=0.14.2",
    "fastapi[standard]>=0.115.13",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
